Made rating service test very robust

+++ backend/src/main/java/com/project/G1_T3/playerprofile/model/PlayerProfile.java
@@ -147,8 +147,8 @@ public class PlayerProfile {
         }
     }
 
-    public void setGlickoRating(int glickoRating) {
-        this.glickoRating = glickoRating;
+    public void setGlickoRating(double glickoRating) {
+        this.glickoRating = (float) glickoRating;
         syncGlicko2Rating();
     }
